pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "doc2mkdocs"
version = "1.0.0"
description = "Convert documentation files (DOCX, PDF, XLSX) into MkDocs-ready Markdown"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["documentation", "mkdocs", "markdown", "converter", "docx", "pdf", "xlsx"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Documentation",
    "Topic :: Text Processing :: Markup :: Markdown",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pymupdf>=1.23.0",
    "openpyxl>=3.1.0",
    "mammoth>=1.6.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "pyyaml>=6.0",
    "python-magic-bin>=0.4.14; platform_system=='Windows'",
    "python-magic>=0.4.27; platform_system!='Windows'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "types-PyYAML",
]
web = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
]
watch = [
    "watchdog>=3.0.0",
]

[project.scripts]
doc2mkdocs = "doc2mkdocs.cli:app"

[project.urls]
Homepage = "https://github.com/yourusername/doc2mkdocs"
Documentation = "https://github.com/yourusername/doc2mkdocs#readme"
Repository = "https://github.com/yourusername/doc2mkdocs"
Issues = "https://github.com/yourusername/doc2mkdocs/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
doc2mkdocs = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/doc2mkdocs --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

